fix:jstd.conf: explicitly specify script load order
Originally we relied on a lot of globbing, which resulted in
angular-mocks being loaded before normal services, so we never overwrote
services like $exceptionHandler with mocks. Explict definition
guarantees that we don't fall into the loading order trap, but requires
us to remember to update the jsTestDriver.conf file every time we
add/rename/remove a js file.